feat : make usePreview
