+++ src/components/atoms/index.ts
@@ -3,3 +3,4 @@ export * from './Modal';
 export * from './HeaderBarItem';
 export * from './Layout';
 export * from './Grid';
+export * from './ImageUpload';
