+++ src/components/molecules/CreateArticleModal/index.tsx
@@ -1,25 +1,13 @@
 import React, { useState } from 'react';
 import { Modal } from 'src/components/atoms';
-import { useS3Upload } from 'next-s3-upload';
-import { compressImage } from 'src/utils';
+import { ImageUpload } from 'src/components/atoms';
 
 export const CreateArticleModal = () => {
   const [imageUrl, setImageUrl] = useState<string>();
 
-  const { FileInput, openFileDialog, uploadToS3 } = useS3Upload();
-
-  const handleFileChange = async (file: File) => {
-    const compressedImage = await compressImage(file);
-    const { url } = await uploadToS3(compressedImage);
-    setImageUrl(url);
-  };
-
   return (
     <Modal>
-      <div>
-        <FileInput onChange={handleFileChange} />
-        <button onClick={openFileDialog}>Upload file</button>
-      </div>
+      <ImageUpload setImageUrl={setImageUrl} />
     </Modal>
   );
 };
