+++ src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './axiosConfig';
 export * from './imageHandler';
 export * from './apiHandler';
+export * from './compressImage';
